make bootstrapping from a jar_library possible!!!
use the new `test-thrift-gen-classpath` goal!!!
