+++ src/python/pants/backend/jvm/rules/hermetic_dist.py
@@ -2,6 +2,7 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 from dataclasses import dataclass
+from typing import cast
 
 from pants.backend.jvm.tasks.jvm_compile.jvm_compile import JvmCompile
 from pants.engine.rules import rule
@@ -13,7 +14,7 @@ class HermeticDist:
 
   @property
   def underlying_home(self) -> str:
-    return self.underlying.underlying_home
+    return cast(str, self.underlying.underlying_home)
 
 
 @rule
